test: consolidate FSM builder coverage (#754)
## Summary

- consolidate the 14 EBNF exact-layout cases into the FSM
sequence-builder test suite added by #727
- remove two JSON Schema snapshots and one structural-tag snapshot whose
layouts change with unrelated converter implementations
- fix the long-sequence helper's Python 3.8-compatible tuple return
annotation

## Test plan

- [x] Build the native library from the latest `main`
- [x] Run all 38 consolidated FSM builder tests with that branch-native
library
- [x] Run Black, isort, and Ruff checks
- [x] Run repository pre-commit hooks
